Java conversion progress
