+++ src/vekta/PauseMenu.java
@@ -1,3 +1,7 @@
+package vekta;
+
+import static vekta.Vekta.*;
+
 /**
   Pause menu implementation as a Context. Eventually this can be converted over to use a Menu for additional flexibility.
 */
@@ -5,8 +9,8 @@ class PauseMenu implements Context {
   private final World world;
   
   // Pause menu options
-  String[] pauseMenu = {"Continue" , "Restart", "Quit to Menu"};
-  int selected = 0;
+  private final String[] pauseMenu = {"Continue" , "Restart", "Quit to Menu"};
+  private int selected = 0;
   
   public PauseMenu(World world) {
     this.world = world;
@@ -14,16 +18,16 @@ class PauseMenu implements Context {
   
   @Override
   void render() {
-    // TODO: render parent context below
+    Vekta v = Vekta.getInstance();
     
     // Border box
-    rectMode(CORNER);
-    stroke(UI_COLOR);
-    fill(0);
-    rect(-1, -1, width / 4, height + 2);
+    v.rectMode(CORNER);
+    v.stroke(UI_COLOR);
+    v.fill(0);
+    v.rect(-1, -1, v.width / 4, v.height + 2);
     // Logo
-    shapeMode(CENTER);
-    shape(logo, width / 8, 100, (width / 4) - 100, ((width / 4) - 100) / 3.392);
+    v.shapeMode(CENTER);
+    v.shape(logo, width / 8, 100, (width / 4) - 100, ((width / 4) - 100) / 3.392);
     // Options
     for(int i = 0; i < pauseMenu.length; i++) {
       drawOption(pauseMenu[i], (height / 2) + (i * 100), i == selected);
@@ -56,16 +60,16 @@ class PauseMenu implements Context {
       if(getSetting("music") > 0) theme.stop();
       // Play the sound for selection
       if(getSetting("sound") > 0) select.play();
+      clearOverlay();
       switch(selected) {
         case(0):
-          unpause();
+          setContext(world);
           break;
         case(1):
           world.restart();
           break;
         case(2):
-          clearContexts();
-          openContext(mainMenu);
+          setContext(mainMenu);
           break;
       }
     }
@@ -76,10 +80,4 @@ class PauseMenu implements Context {
   
   @Override
   public void mouseWheel(int amount) {}
-  
-  private void unpause() {
-    clearOverlay();
-    closeContext(this);
-    openContext(world);
-  }
 }
