+++ src/vekta/World.java
@@ -1,3 +1,4 @@
+package vekta;
 
 interface World extends Context {
   /**
